requirements.txt
unstructured[all-docs]
xata
python-dotenv
pinecone-client
tiktoken
tenacity
pinecone-client
beautifulsoup4
requests
langchain
langchain-pinecone
fastapi
uvicorn[standard]
transformers
trl
peft
accelerate
datasets
bitsandbytes
scipy
sentencepiece
gradio
protobuf
streamlit
nicegui
dashscope
pypinyin